requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
Werkzeug==2.3.7
click==8.1.7
watchdog==3.0.0 